pyproject.toml
[tool.poetry]
name = "depth-estimation"
version = "0.1.0"
description = ""
authors = ["Artem Trybushenko <artem.trybushenko@gmail.com>"]
readme = "README.md"
packages = [
    { include = "depth_estimation", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.10.0.84"
psutil = "^6.1.0"
statistics = "^1.0.3.5"
GPUtil = "^1.4.0"
pandas = "^2.2.3"
seaborn = "^0.13.2"
torch = "^2.5.1"
torchvision = "^0.20.1"
matplotlib = "^3.9.2"
tqdm = "^4.67.0"
ipykernel = "^6.29.5"
pillow = "<11.0"
numpy = "<2.0"
gradio = "4.29.0"
gradio-imageslider = "^0.0.20"
fastapi = {extras = ["standard"], version = "^0.115.5"}
uvicorn = "^0.32.0"
python-multipart = "^0.0.17"
loguru = "^0.7.2"
fastapi-cors = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
